### 2022-12-12 #118 @whedon
[PRE REVIEW]: Julia for HPC: In Situ data Analysis with Julia for Climate Simulations at Large Scale
**Submitting author:** @ltang85 (<a href="http://orcid.org/0000-0002-7636-0876">Li Tang</a>)
**Repository:** <a href="https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale" target ="_blank">https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale</a>
**Version:**
