### 2023-01-02 #118 @carstenbauer
@ltang85 there seems to be an issue with `algorithm2e`. Please try to fix this. You can retrigger compilation any time with the command `@whedon generate pdf` (see above).
