+++ paper/bib.tex
@@ -1,4 +1,4 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
 \bibliographystyle{juliacon}
-\bibliography{ref.bib}
+\bibliography{/home/ltang/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale/paper/ref.bib}
