### 2023-03-20 #118 @ltang85
@carstenbauer I solved several issues and can generate the pdf locally with "latexmk -bibtex -pdf paper.tex". However, I am still getting errors with whedon. Do you know how to access paper.log with whedon? Or can we just submit the generated pdf? Thansk.
