### 2023-09-07 #118 @luraess
Hi @ltang85. I am reaching out as we had a transition period at JuliaCon Proceedings and I am the new editor handling your submission. According to the thread, you encountered some issues wrt generating the PDF for your manuscript. Before we continue, maybe we should sync on 2 important aspects:
1. Are you still interested in getting your work published here
2. Could you resolve the PDF generation issue in the meantime
