+++ paper/ref.bib
@@ -129,6 +129,7 @@ doi = {},
 url = {https://www.osti.gov/biblio/5936274},
 place = {United States},
 year = {1987},
+publisher={Academic press},
 month = {1}
 } 
 
@@ -141,7 +142,7 @@ month = {1}
   organization={IEEE}
 }
 
-@incollection{kelly2020chimbuko,
+@inproceedings{kelly2020chimbuko,
   title={Chimbuko: A workflow-level scalable performance trace analysis tool},
   author={Kelly, Christopher and Ha, Sungsoo and Huck, Kevin and Van Dam, Hubertus and Pouchard, Line and Matyasfalvi, Gyorgy and Tang, Li and D'Imperio, Nicholas and Xu, Wei and Yoo, Shinjae and others},
   booktitle={ISAV'20 In Situ Infrastructures for Enabling Extreme-Scale Analysis and Visualization},
@@ -264,23 +265,6 @@ key = {SENSEI},
   organization={Springer}
 }
 
-@article{bezanson2012julia,
-  title={Julia: A fast dynamic language for technical computing},
-  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
-  journal={arXiv preprint arXiv:1209.5145},
-  year={2012}
-}
-
-@article{aycock2003brief,
-  title={A brief history of just-in-time},
-  author={Aycock, John},
-  journal={ACM Computing Surveys (CSUR)},
-  volume={35},
-  number={2},
-  pages={97--113},
-  year={2003},
-  publisher={ACM New York, NY, USA}
-}
 
 @inproceedings{lattner2004llvm,
   title={LLVM: A compilation framework for lifelong program analysis \& transformation},
@@ -291,16 +275,6 @@ key = {SENSEI},
   organization={IEEE}
 }
 
-@article{golaz2019doe,
-  title={The DOE E3SM coupled model version 1: Overview and evaluation at standard resolution},
-  author={Golaz, Jean-Christophe and Caldwell, Peter M and Van Roekel, Luke P and Petersen, Mark R and Tang, Qi and Wolfe, Jonathan D and Abeshu, Guta and Anantharaj, Valentine and Asay-Davis, Xylar S and Bader, David C and others},
-  journal={Journal of Advances in Modeling Earth Systems},
-  volume={11},
-  number={7},
-  pages={2089--2129},
-  year={2019},
-  publisher={Wiley Online Library}
-}
 
 @article{thompson2018decline,
   title={The Decline of Computers as a General Purpose Technology: Why Deep Learning and the End of {M}oore's {L}aw are Fragmenting Computing},
