### 2023-09-17 #118 @luraess
Hi @ltang85, I was wondering if you could make any progress on the PDF generation issue? Ideally, we have this resolved within the coming week else we cannot proceed with the review process. Please try to get the PDF generation to work by September 22. Thanks
