+++ paper/paper.tex
@@ -4,8 +4,8 @@
 \setcounter{page}{1}
 
 \usepackage{amsmath}
-%\usepackage[ruled,vlined]{algorithm2e}
-\usepackage{algorithm, algorithmic}
+\usepackage[ruled,vlined]{algorithm2e}
+%\usepackage{algorithm, algorithmic}
 
 \usepackage{booktabs}
 \usepackage[width=.45\textwidth]{caption}
