+++ paper/test.tex
@@ -1,0 +1,426 @@
+
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+
+\usepackage{amsmath}
+%\usepackage[ruled,vlined]{algorithm2e}
+%\usepackage{algorithm, algorithmic}
+
+\usepackage{booktabs}
+\usepackage[width=.45\textwidth]{caption}
+
+\usepackage{graphicx}
+\graphicspath{{../figures/}}
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+Fast-evolving data science techniques have enabled scientific discoveries in part by providing the ability to analyze large amounts of scientific data and extract relevant patterns.  Scientific simulations offer a unique opportunity to study large-scale, complex systems, such as the Earth's climate, under varying conditions and modeling assumptions. However, one major obstacle to improved utilization of large scientific simulations is the ever-increasing gap between computing speed, which continues to increase, and data I/O bandwidth, which remains relatively constant. This mismatch often prevents full utilization of the data; we are likely unable to save all of the generated data for analysis, and simple solutions such as saving compressed versions of the data for later analysis may be insufficient to answer questions of interest.
+
+In situ data analysis techniques seek to address this issue by completing analysis and pattern extraction from generated scientific data while the application is running, and the in situ data analysis only involves compute and in-memory data I/O, which improves performance. However, one major challenge of employing in situ data analysis to legacy scientific applications is that modern advanced data science techniques are usually not implemented in the same way as the scientific codes (typically, using the Fortran programming language with parallel programming models). To address this challenge, we develop an infrastructure for coupling a popular high-level data science programming language, Julia, with the large-scale production-level climate code Energy Exascale Earth System Model (E3SM) on high performance computing (HPC) systems. To demonstrate the infrastructure, we develop two in situ data analysis methods in Julia and evaluate their performance and the infrastructure overhead. Our results show that our in situ Julia data analysis methods are able to detect extreme weather events and characterize climate patterns with insignificant infrastructure overhead. Furthermore, our infrastructure allows user-friendly development and deployment of new Julia data analysis modules without the need to recompile the simulation code, giving data analysts a simple new tool for in situ analysis.
+
+%\headingtable
+
+\end{abstract}
+
+\section{Introduction}
+
+Computing devices are evolving much faster than storage devices because of their different ratios of performance improvement to R\&D investment in the post-Moore’s Law era~\cite{thompson2018decline}. This trend enlarges the performance gap between our computing and storage devices and greatly limits our ability to analyze all the generated scientific simulation data. When run on HPC systems, modern physics simulations employ thousands of nodes, generating large amounts of data. While this data could possibly be stored for post-processing, write time to disk, available storage space, and post-processing time can be prohibitive, especially at high physics simulation resolution.  At such a high resolution, the three-dimensional physics model state alone can be tens of gigabytes \emph{per} time slice. To address these issues, we adopt in situ data analysis to analyze physics simulation data as the simulation is running. 
+
+
[…]
