### 2023-09-18 026cd8e
Delete paper/test.tex
