### 2023-09-18 ad1f96c
Update paper.tex
