### 2023-09-18 a4ecf1c
Update paper.tex
