+++ paper/paper.tex
@@ -210,13 +210,16 @@ Here, we briefly present the distributed SSW detection technique that we have de
 \Statex
 \For{each EAM time\_step}
   \If{(current\_time\_step == end\_of\_day)}
-    \State Compute partial zonal mean zonal wind per MPI rank.
-    \State Compute the global zonal mean zonal wind.
+    \State Compute partial zonal mean zonal wind values
+    \State at each MPI process.
+    \State Compute the global zonal mean zonal wind
+    \State values with MPI:Reduction.
     \If{(global zonal mean zonal wind < 0)}
     \State counter=counter+1
     \EndIf
     \If{((counter == 10))}
-    \State Record current time step as the beginning of a SSW event.
+    \State Record current time step as the beginning of
+    \State a SSW event.
     \State counter=0
     \EndIf
   \Else
