### 2023-09-18 #118 @luraess
> Hi @luraess, I have fixed the algorithm2e issue. Please let me know if there is still any issue with the pdf generation. Thanks.

Hi @ltang85 , thanks for fixing this! We will be able now to proceed with starting the revision.
