### 2023-10-12 #134 @whedon
[REVIEW]: Julia for HPC: In Situ data Analysis with Julia for Climate Simulations at Large Scale
**Submitting author:** @ltang85 (<a href="http://orcid.org/0000-0002-7636-0876">Li Tang</a>)
**Repository:** https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
