### 2023-11-08 #134 @simonbyrne
This was an interesting application of embedding Julia inside another program for in situ analysis. Although I am not able to assess the scientific validity of the analysis (which seems incomplete), from a technical perspective, this is an important problem: as simulations reach higher resolutions, it becomes more expensive and difficult to store all the data for post-processing analysis, thus it is helpful if such analysis can be done while the simulation is running ("in situ").

Overall, I thought the paper was reasonable, but the exposition could be improved in a few places.

There is no linked code, (other than the TributaryPCA.jl package, and this doesn't meet most of the above criteria, such as documentation), so I can't really check off those relevant boxes. While the code may not strictly be required, it is a shame as I think this could be of interest to others who are attempting something similar. I would encourage the authors to make those available if possible.

# Specific comments
## 3.3 C Interface

> It is also important to conduct garbage collection
(GC) in this C interface for every time step. This is because arr
and p_arr in the code snippet are allocated and then deallocated
in E3SM for every time step, and the Julia runtime is not aware of
this and hence needs explicit GC

This is not what the C code in code 2 is doing: as described in https://docs.julialang.org/en/v1/manual/embedding/#Memory-Management, the `JL_GC_PUSH3` macro “roots” the variables so they are _not_ collected, and `JL_GC_POP` simply removes that root (so they can be cleaned up by the garbage collector). The Julia garbage collector runs automatically (unless it is disabled, but this is not described in the paper)

## 3.5 Noninvasive Compilation
Some more details could be useful here: how did you load the Julia code? (e.g. did you make use of a precompiled system image?) How did you switch between (or switch off) the in situ analysis codes? Again, having the actual code available could make this easier.

## 4.1 Sudden Stratospheric Warming (SSW)

I'm not sure Algorithm 1 is correct: shouldn't it reset the `counter` to 0 if "global zonal mean zonal wind >= 0"? If it is negative for 20 days, does that count as 2 events?

There isn't much detail on the actual GEV fitting, and I found it difficult to understand. Is it fitting a different GEV distribution at every point? What is the prior distribution used for the GEV parameters? Some more details on this would be helpful (as well as linking to the code).

## 4.2 TributaryPCA

While I understand the problem it is trying to solve, I found it difficult to understand algorithm, and the Code 4 block is somewhat unclear:
- What is `X_par`? Is it a matrix or a vector, and what are the dimensions?
- What is the initial value (and size) of alpha and `V_par`?
- Some lines are written using in-place allocation free functions (e.g. `mul!`), yet not others. There are several calls to `copy` which seem completely unnecessary.

(I don't have access to the TributaryPCA paper, so couldn't look it up there either).

I would suggest either having a mathematical exposition of the algorithm, and/or making the code a bit more readable.

Is the data centered (i.e. do you subtract the mean)? The lack of red in Fig 3, PC1 suggests that it isn't, but that isn't clear. If you aren't then the interpretation changes slightly.

## 5.2 In Situ SSW Detection and Characterization Results

I don't quite understand this: it doesn't actually analyse the SSW case? It just seems to be an illustration of how one might do an analysis.

## 5.3 PCA Results

The details are a bit lacking. Was the data collected every time step? Or is it based on daily snapshots or averages? How long was the simulation?

I disagree with the interpretation of figure 3: you can't really assign "warmer vs colder" to principal components (the sign of eigenvectors doesn't matter). My interpretation of the figures is that:
- PC1: the lack of red suggests the data hasn't been centered (or is centered around the equatorial temperature?). As a results it is just explaining that "poles are colder than the equator"
- PC2 and PC3: these look to be roughly the first Fourier components by longitude, i.e. are some orthogonal variation of `sin(long)` and `cos(long)`: my guess is that they're just capturing the daily temperature changes as the sun rotates around the earth? Can you plot the corresponding eigenvalues?

While it's useful that you can capture this, it's not particularly scientifically interesting.

## 5.4 Performance

The SSW seems to have negligible overhead (which is not surprising since it is evaluated infrequently), so there is not much to say about the analysis.

It could be helpful if you could add the number of MPI ranks to Table 1?

I'd be interested to know more about the performance bottlenecks of the PCA code? Which parts were the most expensive? Did the Julia garbage collector cause problems with scaling?

# Minor comments

- Should “in situ” be italicized? I don't know if there is a style guide.
- The formatting of the references is somewhat messed up (e.g. many abbreviations are uncaptialized), and many DOIs are missing (e.g. the TributaryPCA one).
- Since you make use of it, it would be nice if you could cite the MPI.jl paper (https://github.com/JuliaParallel/MPI.jl#citation) [disclaimer: I’m one of the authors, but it is the only software mentioned which is uncited]
- Slurm appears both capitalized and uncapitalized.
