+++ paper/paper.tex
@@ -293,7 +293,7 @@ Grizzly uses Slurm for job scheduling and resource management. The total storage
 To deploy our in situ algorithms, we have run the fully coupled E3SM simulation and accessed the data from its Atmosphere model, EAM, for analyses. The EAM module runs in conjunction with other coupled modules (MPAS-Ocean, MPAS Sea Ice, MPAS Land Ice, and Land Surface) to produce scientifically meaningful climate data. In this work, we use a realization of the Shared Socioeconomic Pathway (SSP) $585$ scenario~\cite{Eyring} with E3SM. This is an aggressive scenario that assumes the climate will experience an increase in radiative forcing of $8.5$ W/m$^2$. We use the standard E3SM V1 configuration with a $1^{\circ}$ atmosphere and land (equivalent to 110 km at the equator), $0.5^{\circ}$ river model ($55$ km), and an ocean and sea ice with mesh spacing varying between $60$ km in the midlatitudes and $30$ km at the equator and poles~\cite{e3sm_v1}. %To run optimally, E3SM simulation test cases are often configured to run on a specific number of MPI processes and this particular case is optimized to run on $3024$ cores, which is equivalent to running it on $84$ computing nodes in Grizzly.
 
 
-Our tests fall into two data analysis tasks: SSW and PCA. Importantly, E3SM is only compiled once, and we can then switch between the PCA and SSW in situ Julia modules. The Julia modules are dynamically compiled and executed by using the Julia $1.8$ version. Each in situ run is performed 10 times to mitigate system noise. We also add timers in the slurm system for measuring performance and overhead.
+Our tests fall into two data analysis tasks: SSW and PCA. Importantly, E3SM is only compiled once, and we can then switch between the PCA and SSW in situ Julia modules. The Julia modules are dynamically compiled and executed by using the Julia $1.8$ version. Each in situ run is performed 10 times to mitigate system noise. We also add timers in the Slurm system for measuring performance and overhead.
 
 \subsection{In Situ SSW Detection and Characterization Results}
 After running for 365 simulation days (one simulated year), with the GEV models updated every three days, we found that the post-SSW GEV model was updated 11 total times while the non-SSW GEV model was updated 110 times.
