+++ src/E3SMFiles/stepon.F90
@@ -1,0 +1,736 @@
+!BOP
+!
+! !MODULE: stepon -- FV Dynamics specific time-stepping
+!
+! !INTERFACE:
+module stepon
+
+! !USES:
+! from cam
+   use shr_kind_mod,   only: r8 => shr_kind_r8
+   use shr_sys_mod,    only: shr_sys_flush
+   use pmgrid,         only: plev, plevp, plat
+   use spmd_utils,     only: iam, masterproc, mpicom, npes
+   use constituents,   only: pcnst, cnst_name, cnst_longname
+   use cam_abortutils, only: endrun
+   use ppgrid,         only: begchunk, endchunk
+   use physconst,      only: zvir, cappa
+   use physics_types,  only: physics_state, physics_tend
+   use dyn_comp,       only: dyn_import_t, dyn_export_t
+   use perf_mod,       only: t_startf, t_stopf, t_barrierf
+   use time_manager,   only: get_step_size
+! from SE
+   use derivative_mod, only: derivinit, derivative_t
+   use viscosity_mod, only : compute_zeta_C0, compute_div_C0
+   use quadrature_mod, only: gauss, gausslobatto, quadrature_t
+   use edge_mod,       only: edge_g, edgeVpack_nlyr, edgeVunpack_nlyr
+   use parallel_mod,   only : par
+   use scamMod,        only: use_iop, doiopupdate, single_column, &
+                             setiopupdate, readiopdata
+   use element_mod,    only: element_t
+   use shr_const_mod,       only: SHR_CONST_PI
+
+   implicit none
+   private
+   save
+
+!
+! !PUBLIC MEMBER FUNCTIONS: 
+!
+  public stepon_init   ! Initialization
+  public stepon_run1    ! run method phase 1
+  public stepon_run2    ! run method phase 2
+  public stepon_run3    ! run method phase 3
+  public stepon_final  ! Finalization
+
+  public insitu_processing ! insitu subroutine !  
+
+!----------------------------------------------------------------------
+!
+! !DESCRIPTION: Module for dynamics time-stepping.
+!
+! !REVISION HISTORY:
+!
+! 2006.05.31  JPE    Created
+! 2019.06.28  MT     Updated to output vorticity/divergence, new DSS interface
+!
+!EOP
+!----------------------------------------------------------------------
+!BOC
+!
+! !PRIVATE DATA MEMBERS:
+!
+
+  logical :: iop_update_phase1
+
+  type (derivative_t)   :: deriv           ! derivative struct
+  type (quadrature_t)   :: gv,gp           ! quadratures on velocity and pressure grids
+!-----------------------------------------------------------------------
+
+
+CONTAINS
+
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+!----------------------------------------------------------------------- 
+!BOP
+! !ROUTINE:  stepon_init --- Time stepping initialization
+!
+! !INTERFACE:
+subroutine stepon_init(dyn_in, dyn_out )
+! !USES:
+  use dimensions_mod,         only: nlev, nelemd, npsq
+  use dyn_grid,               only: fv_nphys
+  use cam_history,            only: addfld, add_default, horiz_only
+  use cam_history,            only: register_vector_field
[…]
