+++ src/README.md
@@ -1,214 +1,184 @@
-This documentation shows how to run the E3SM InSitu framework with the streaming Julia module on LANL's Grizzly.
-
-Contact: Li Tang (ltang@lanl.gov)
-
-## STEP 1. Install and run E3SM example on Grizzly
-=====================================================================
-This is work in progress and will be updated as needed to keep it up to date.
-The document follows the slides provided by Luke Van Roekel and has modifications.
-Created by: Soumya Dutta (sdutta@lanl.gov)
-export PROJECT varibale to make sure that you are using the correct LANL IC allocation
-export PROJECT=w20_insituinf
-Get the repository: (make sure to add ssh key on github before you clone the repository)
-git clone git@github.com:E3SM-Project/E3SM.git
-cd E3SM
-git fetch origin
-git reset --hard origin/vanroekel/lanl_dr
-git submodule update --init
-Setup a new case:
-cd cime/scripts
-Now create the new case. In this step you can specify the model you want to run, what resolution
-of mesh you want and the compiler that will be used to build and run.
-./create_newcase -case /users/sdutta/climate/e3sm_case_1 -compset FC5AV1C-L -mach grizzly -
-res ne4_ne4 -compiler gnu --pecount S
-Here e3sm_case_1 is the name of the directory I gave. It can be anything where all the case
-specific files will be produced necessary for bulding and running.
-open file: env_run.xml
-set value for DIN_LOC_ROOT = /lustre/scratch3/turquoise/lvanroekel/E3SM/input_data (line
-885)
-change STOP_OPTION to nmonths (this sets the unit for STOP_OPTION to months)
-STOP_N to 5 (this specifies that we are looking to run the model for 5 months) save and exit:
-env_run.xml
-open file: user_nl_cam
-add the following:
-avgflag_pertape = 'A','A','I','A','A'
-fincl1 = 'extinct_sw_inp','extinct_lw_bnd7','extinct_lw_inp'
-fincl2 = 'FLUT', 'PRECT', 'U200', 'V200', 'U850', 'V850', 'Z500', 'OMEGA500', 'UBOT', 'VBOT',
-'TREFHT', 'TREFHTMN', 'TREFHTMX', 'QREFHT', 'TS', 'PS', 'TMQ', 'TUQ', 'TVQ'
-fincl3 = 'PSL','T200','T500','U850','V850','UBOT','VBOT','TREFHT'
-fincl4 = 'FLUT','U200','U850','PRECT','OMEGA500'
-fincl5 = 'PRECT','PRECC'
-mfilt = 1,30,120,120,240
-nhtfrq = 0,-24,-6,-6,-3
-save and exit: user_nl_cam
-setup and build the case:
-./case.setup
-./case.build
-open file: env_workflow.xml
-under case.run section set the value of JOB_WALLCLOCK_TIME = 00:25:00.
-JOB_WALLCLOCK_TIME determines how long you want to run your job. save and exit:
-env_workflow.xml
-Run the case:
-./case.submit
-The process will create a directory called E3SM/e3sm_case_1 in your scratch space.
-For me the path is: /lustre/scratch3/turquoise/sdutta/E3SM/e3sm_case_1
-Check job status: squeue -u $USER
-
[…]
