+++ paper/ref.bib
@@ -6,6 +6,7 @@
   volume={59},
   number={1},
   pages={65--98},
+  doi={https://doi.org/10.48550/arXiv.1209.5145},
   year={2017},
   publisher={SIAM}
 }
@@ -15,51 +16,76 @@
   author={Ge, Hong and Xu, Kai and Ghahramani, Zoubin},
   booktitle={International conference on artificial intelligence and statistics},
   pages={1682--1690},
+  doi={https://doi.org/10.17863/CAM.42246},
   year={2018},
   organization={PMLR}
 }
 
+@article{godoy2020adios,
+  title={{ADIOS} 2: The adaptable input output system. a framework for high-performance data management},
+  author={Godoy, William F and Podhorszki, Norbert and Wang, Ruonan and Atkins, Chuck and Eisenhauer, Greg and Gu, Junmin and Davis, Philip and Choi, Jong and Germaschewski, Kai and Huck, Kevin and others},
+  journal={SoftwareX},
+  volume={12},
+  pages={100561},
+  doi={https://doi.org/10.1016/j.softx.2020.100561},
+  year={2020},
+  publisher={Elsevier}
+}
+
+@inproceedings{byrne2021mpi,
+  title={{MPI}.jl: Julia bindings for the Message Passing Interface},
+  author={Byrne, Simon and Wilcox, Lucas C and Churavy, Valentin},
+  booktitle={Proceedings of the JuliaCon Conferences},
+  volume={1},
+  number={1},
+  pages={68},
+  year={2021}
+}
 
 @article{xie2018understanding,
-  title={Understanding cloud and convective characteristics in version 1 of the E3SM atmosphere model},
+  title={Understanding cloud and convective characteristics in version 1 of the {E3SM} atmosphere model},
   author={Xie, Shaocheng and Lin, Wuyin and Rasch, Philip J and Ma, Po-Lun and Neale, Richard and Larson, Vincent E and Qian, Yun and Bogenschutz, Peter A and Caldwell, Peter and Cameron-Smith, Philip and others},
   journal={Journal of Advances in Modeling Earth Systems},
   volume={10},
   number={10},
   pages={2618--2644},
+  doi={https://doi.org/10.1029/2018MS001562},
   year={2018},
   publisher={Wiley Online Library}
 }
 
 @article{golaz2019doe,
-  title={The DOE E3SM coupled model version 1: Overview and evaluation at standard resolution},
+  title={The {DOE E3SM} coupled model version 1: Overview and evaluation at standard resolution},
   author={Golaz, Jean-Christophe and Caldwell, Peter M and Van Roekel, Luke P and Petersen, Mark R and Tang, Qi and Wolfe, Jonathan D and Abeshu, Guta and Anantharaj, Valentine and Asay-Davis, Xylar S and Bader, David C and others},
   journal={Journal of Advances in Modeling Earth Systems},
   volume={11},
   number={7},
   pages={2089--2129},
+  doi={https://doi.org/10.1029/2018MS001603},
   year={2019},
   publisher={Wiley Online Library}
 }
 
 @article{qian2018parametric,
-  title={Parametric sensitivity and uncertainty quantification in the version 1 of E3SM atmosphere model based on short perturbed parameter ensemble simulations},
+  title={Parametric sensitivity and uncertainty quantification in the version 1 of {E3SM} atmosphere model based on short perturbed parameter ensemble simulations},
[…]
