+++ paper/paper.tex
@@ -418,7 +418,7 @@ We first run the high-resolution E3SM problem with both of the SSW\_GEV and PCA 
 
 \begin{figure}
     \centering
-    \includegraphics[width=\linewidth, height=5.5cm]{paper/figures/analysisoverhead.pdf}
+    \includegraphics[width=\linewidth, height=5.5cm]{figures/analysisoverhead.pdf}
     \caption{In situ data analysis overhead (ratios of in situ data analysis time to the total EAM time).}
     \label{fig:ssw_break}
 \end{figure}
@@ -426,7 +426,7 @@ We first run the high-resolution E3SM problem with both of the SSW\_GEV and PCA 
 
 \begin{figure}
     \centering
-    \includegraphics[width=\linewidth, height=5.5cm]{paper/figures/infrastructureoverhead.pdf}
+    \includegraphics[width=\linewidth, height=5.5cm]{figures/infrastructureoverhead.pdf}
     \caption{In situ infrastructure overhead (ratios of in situ infrastructure overhead time to the total EAM time).}
     \label{fig:pca_break}
 \end{figure}
