+++ paper/paper.tex
@@ -62,446 +62,7 @@ This paper is organized in the following structure. Section~\ref{sec2} provides 
 
 
 
-\section{Background and Related Work} \label{sec2}
 
-In this section, we describe the E3SM HPC simulation, the Julia programming language, and the general concept of in situ data analysis before discussing recent closely related work.
-
-
-
-
-
-Our work aims to provide data scientists with a performant interface for developing in situ data analysis algorithms without directly interacting with complex HPC codes or relying on post-hoc analysis. However, there are three challenges in doing so. First, E3SM's EAM is written in Fortran and adopts Message Passing Interface (MPI) for large-scale parallelism~\cite{walker1996mpi}, and there are no Fortran APIs for embedding Julia into EAM. Second, one MPI rank of EAM is only able to access to its local data (e.g., velocity and temperature), and each in situ Julia data analysis instance can only access the same local data of its parent EAM MPI rank. Last, EAM is compiled and interacting with other E3SM components, and the in situ Julia data analysis should be noninvasive to other E3SM components. To address these challenges, we develop and implement an infrastructure for coupling the Julia runtime with E3SM, which enables effective transfer of data and consistent MPI communication between Julia and E3SM through a minimalist design. 
-
-
-\subsection{E3SM}
-
-E3SM is the new earth system model developed by DOE to answer critical DOE mission questions relating to energy and water security \emph{and} run efficiently on emerging HPC architectures.  E3SM includes components to model the atmosphere, land, rivers, ocean, and sea ice. The atmosphere component, EAM, utilizes a spectral element discretization on a cubed sphere grid.  The model uses approximately 110 km resolution in the horizontal and 72 vertical layers with variable spacing.  The model top is at 60km above ground level.  A detailed analysis of its performance is given in~\cite{xie2018understanding,zhang2019evaluation,qian2018parametric,rasch2019overview,golaz2019doe}.  EAM has a number of key features relative to its predecessor, which are described in~\cite{golaz2019doe}.  Of interest to our work is the linearized ozone photochemistry, which improves EAM simulations of the stratosphere, which is critically important for our SSW analysis.
-
-
-
-\subsection{Julia}
-
-Julia is a dynamically-typed programming language that is for general-purpose application development and provides many Python-like features (e.g., rich package support)~\cite{bezanson2012julia}. The first public Julia release was introduced in 2013 and it is designed to provide users with the speed of C/C++/Fortran~\cite{bezanson2017julia} while giving a convenient high-level programming environment. Julia is a built on top of the modularized LLVM compiler infrastructure~\cite{lattner2004llvm}, and features a Just-In-Time (JIT) compilation engine~\cite{aycock2003brief}. Therefore, Julia is ideal for agile development of advanced data science algorithms due to its speed and effectiveness in solving computationally complex problems. Also, Julia has support for parallelism and concurrency, which are crucial for HPC applications. In this paper, our major goal is to increase productivity by allowing data scientists to focus on writing high-level, performant numerical algorithms for legacy HPC applications at large-scale.
-
-
-
-
-
-
-\subsection{In Situ Data Analysis} 
-
-
[…]
