+++ paper/ref.bib
@@ -6,7 +6,6 @@
   volume={59},
   number={1},
   pages={65--98},
-  doi={https://doi.org/10.48550/arXiv.1209.5145},
   year={2017},
   publisher={SIAM}
 }
@@ -37,7 +36,6 @@
   author={Byrne, Simon and Wilcox, Lucas C and Churavy, Valentin},
   booktitle={Proceedings of the JuliaCon Conferences},
   volume={1},
-  number={1},
   pages={68},
   year={2021}
 }
@@ -119,13 +117,6 @@
 	year = {1990}
 }
 
-@article{bezanson2012julia,
-  title={Julia: A fast dynamic language for technical computing},
-  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
-  journal={arXiv preprint arXiv:1209.5145},
-  year={2012},
-  doi={https://doi.org/10.48550/arXiv.1209.5145}
-}
 
 @inproceedings{wang2021tributarypca,
   title={Tributary{PCA}: Distributed, Streaming {PCA} for In Situ Dimension Reduction with Application to Space Weather Simulations},
@@ -296,7 +287,7 @@ key = {SENSEI},
   author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
   booktitle={European Parallel Virtual Machine/Message Passing Interface Users’ Group Meeting},
   pages={97--104},
-  doi={https://doi.org/10.1007/978-3-540-30218-6_19},
+  doi={https://doi.org/10.1007/978-3-540-30218-6\_19},
   year={2004},
   organization={Springer}
 }
@@ -542,6 +533,7 @@ publisher={Copernicus GmbH}
 
 @misc{scorpio,
   title = {{SCORPIO}, A high-level Parallel I/O Library},
+  key = {SCORPIO},
   howpublished = {\url{https://github.com/E3SM-Project/scorpio/}}
 }
 
