+++ paper/paper.tex
@@ -103,7 +103,7 @@ The in situ approach has been studied in recent years and has been employed in d
 
 The demand for scalable algorithms for analyzing climate data is growing rapidly as computational climate simulations scale up, leading to growing output data size and prompting scientists to turn to in situ analysis approaches. For example, in situ analysis of eddies in large-scale MPAS-Ocean simulation data was done by Woodring et al.~\cite{7192723}. In this work, we focus on two data analysis questions of interest to climate scientists and demonstrate their in situ application. For extreme weather event analysis, the connection of SSW with the polar vortex events is a focus area for climatologists~\cite{ssw_survey}. More recently, scientists have also used extreme-value distribution models for capturing rare climate phenomena and further analyzing them in detail~\cite{gev_huang2016estimating, gev_wang2017detection}. In our recent work, the usefulness of in situ SSW analysis was also demonstrated~\cite{ssw_isav}.
 For general summarization of large-scale patterns, principal components analysis (PCA) is a standard method that typically requires access to the full simulation data across time steps to capture seasonal trends and deviations.
-TributaryPCA is an adaptation of PCA, a statistical method for reducing the dimensionality of big data, to the in situ setting. Ordinary PCA summarizes variation by projecting high-dimensional data into a lower-dimensional space optimized to retain fundamental properties of the data). TributaryPCA is a distributed version of an existing algorithm for online PCA (Oja's algorithm) that uses the MPI standard for communication. In this way, TributaryPCA is able to estimate the principal components from data that arrives sequentially over time and that is distributed across compute nodes. In previous work from our team, the accuracy and computational efficiency of the TributaryPCA method was demonstrated~\cite{wang2021tributarypca}, but it was not applied in situ. An open-source Julia implementation is available at \url{https://github.com/lanl/PRISM/tree/master/TributaryPCA.jl}.
+TributaryPCA is an adaptation of PCA, a statistical method for reducing the dimensionality of big data, to the in situ setting. Ordinary PCA summarizes variation by projecting high-dimensional data into a lower-dimensional space optimized to retain fundamental properties of the data). TributaryPCA is a distributed version of an existing algorithm for online PCA (Oja's algorithm) that uses the MPI standard for communication. In this way, TributaryPCA is able to estimate the principal components from data that arrives sequentially over time and that is distributed across compute nodes. In previous work from our team, the accuracy and computational efficiency of the TributaryPCA method was demonstrated~\cite{wang2021tributarypca}. An open-source Julia implementation is available at \url{https://github.com/lanl/PRISM/tree/master/TributaryPCA.jl}.
 
 
 
