### 2024-05-07 #134 @luraess
> @luraess I couldn't check the boxes in [#134 (comment)](https://github.com/JuliaCon/proceedings-review/issues/134#issue-1939993021)

I recall now this issue we had already at the beginning - let's leave that out. Could you however briefly assess whether the revised version of the manuscript draft adresses your comments and suggestions @williamfgc ? Thanks!
