### 2024-05-07 #134 @williamfgc
I just spotted typos: e.g. consqeuent...please proof-read. Other than that, my comments were addressed. Thanks!
