+++ paper/paper.tex
@@ -33,18 +33,18 @@ Fast-evolving data science techniques have enabled scientific discoveries by pro
 
 \section{Introduction}
 
-Computing devices are evolving much faster than storage devices because of their different ratios of performance improvement to Research and Development (R\&D) investment in the post-Moore’s Law era~\cite{thompson2018decline}. This trend enlarges the performance gap between our computing and storage devices and greatly limits our ability to analyze all the generated scientific simulation data. When run on High-Performance Computing (HPC) systems, modern physics simulations employ thousands of nodes, generating large amounts of data. For example, at a high resolution, the three-dimensional physics model state alone can be tens of gigabytes \emph{per} time slice. While this data could possibly be compressed and then stored with high-performance input/output (I/O) for post-processing~\cite{jin2022accelerating,scorpio}, data movement and post-processing time can be extremely long. To address these issues, we adopt in situ data analysis to analyze physics simulation data as the simulation is running. The in situ data analysis design involves several design and performance tarde-offs in terms of coupling the physics simulation and data analysis, and we will explore some of the trade-offs in this paper.
+Computing devices are evolving much faster than storage devices because of their different ratios of performance improvement to Research and Development (R\&D) investment in the post-Moore’s Law era~\cite{thompson2018decline}. This trend enlarges the performance gap between our computing and storage devices and greatly limits our ability to analyze all the generated scientific simulation data. When run on High-Performance Computing (HPC) systems, modern physics simulations employ thousands of nodes, generating large amounts of data. For example, at a high resolution, the three-dimensional physics model state alone can be tens of gigabytes \emph{per} time slice. While this data could possibly be compressed and then stored with high-performance input/output (I/O) for post-processing~\cite{jin2022accelerating,scorpio}, data movement and post-processing time can be extremely long. To address these issues, we adopt in situ data analysis to analyze physics simulation data as the simulation is running. The in situ data analysis design involves several design and performance trade-offs in terms of coupling the physics simulation and data analysis, and we will explore some of the trade-offs in this paper.
 
 
 
-Our target HPC application is Energy Exascale Earth System Model (E3SM)~\cite{golaz2019doe}. E3SM is the Department of Energy (DOE) state-of-the-science earth system model, and we focus on E3SM's atmosphere component (EAM) in this study. Developing data science techniques is extremely costly by using HPC programming languages such as C/C++ and Fortran due to the ecosystem lacking, so most data analysis programs are implemented by using Python. However, Python's performance is usually slow on HPC systems and we decided to use Julia as the programming language of developing data analysis modules because of its superior performance for data analysis and rich support of data analysis packages~\cite{hunold2020benchmarking}. Therefore, the major challenge of employing in situ data analysis to legacy scientific applications is that modern advanced data science techniques are usually not implemented in the same way as the HPC scientific codes. In this paper, we hypothesize that the Fortran-based EAM can be effectively coupled with in situ Julia-based data analysis modules on HPC systems.
[…]
