+++ src/README.md
@@ -1,4 +1,4 @@
-This documentation shows how to run the E3SM in situ framework with the streaming Julia modules on LANL's Grizzly.  
+This documentation shows how to run the E3SM in situ framework with the streaming Julia modules on LANL's Grizzly supercomputer.  
   
 Contact: Li Tang (ltang@lanl.gov)  
   
@@ -19,7 +19,7 @@ cd cime/scripts
 Set the path of the input data and caserun infomation (e.g., run the model for 5 months) for the created E3SM case in **env_run.xml**:  
   
 ```
-DIN_LOC_ROOT = **your input_data path**  
+DIN_LOC_ROOT = "your input_data path"  
 STOP_OPTION = nmonths  
 STOP_N = 5
 ```
@@ -49,70 +49,20 @@ Test run:
   
 <h1>STEP 2. Add Julia compilation and linking support to E3SM for Grizzly</h1>  
 
-The E3SM version installed by STEP 1 has two separate systems for compilation and linking.  
-1. The E3SM compilation uses the new CIME CMake system and the changes are made in E3SM/components/CMakeLists.txt  
-IN line 370  
-CHANGE set(INCLDIR "-I.")  
-TO set(INCLDIR "${INCLDIR} -I/turquoise/users/ltang/Install/julia-1.4.0/include/julia")  
-2. The E3SM top linking process uses the old Make system and the changes are made in E3SM/cime/scripts/Tools/Makefile  
-IN Line 956  
-APPEND \$(JULIALIBS) to the end of linking the command line  
-Add JULIALIBS = -L'/turquoise/users/ltang/Install/julia-1.4.0/lib' -Wl,--export-dynamic -Wl,-rpath,'/turquoise/users/ltang/Install/julia-1.4.0/lib' -Wl,-rpath,'/turquoise/users/ltang/Install/julia1.4.0/lib/julia' -ljulia  
-TO (Note: /turquoise/users/ltang/Install/julia-1.4.0 is my Julia installation path, please change it to your Julia installation path)  
-4. Change the software versions for Grizzly in cime/config/e3sm/machines/config_machines.xml, try to find the following configuration in config_machines.xml and then only make changes in the RED lines. You might also need to delete <modules><command name="load">parallel-netcdf/1.5.0</command></modules>. It works with parallel-netcdf/1.5.0  
-<machine MACH="grizzly">  
-<DESC>LANL Linux Cluster, 36 pes/node, batch system slurm</DESC>  
-<NODENAME_REGEX>gr-fe.*.lanl.gov</NODENAME_REGEX>  
-<OS>LINUX</OS>  
-<COMPILERS>gnu,intel</COMPILERS>  
-<MPILIBS>mvapich,openmpi</MPILIBS>  
-<PROJECT>e3sm</PROJECT>  
-<CIME_OUTPUT_ROOT>/lustre/scratch3/turquoise/$ENV{USER}/E3SM/scratch</CIME_OUTPUT_ROOT>  
-<DIN_LOC_ROOT>/lustre/scratch3/turquoise/$ENV{USER}/E3SM/input_data</DIN_LOC_ROOT>  
-<DIN_LOC_ROOT_CLMFORC>/lustre/scratch3/turquoise/$ENV{USER}/E3SM/input_data/atm/datm7</DIN_LOC_ROOT_CLMFORC>  
-<DOUT_S_ROOT>/lustre/scratch3/turquoise/$ENV{USER}/E3SM/archive/$CASE</DOUT_S_ROOT>  
[…]
