### 2024-05-24 #134 @ltang85
@luraess @simonbyrne Sorry for my late reply.

@luraess I have cleaned up the README.md file using the standard markdown syntax. Please check and let me know if you have any comments.

@simonbyrne We have (1) added the code and instructions for reproducibility, (2) corrected the mistakes in the SSW algorithm, (3) added more description texts for SSW and  TributaryPCA, (4) made edits to address your other comments such as results, tables, and garbage collection. More details can be found in our response letter to your comments: https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale/blob/main/paper/Response%20Letter.pdf
