### 2024-05-24 #134 @simonbyrne
Thanks @ltang85, I has missed the response.

In section 5.3, how long was the simulation? It does still look to me like PC2 & PC3 corresponds to the regular diurnal cycle (which I would assume is the biggest source of regular variation in the data)
