+++ paper/paper.tex
@@ -368,6 +368,7 @@ Each component indicates a different spatial pattern of warming (red color) and 
 In particular, the first component indicates a temperature differential between equatorial and polar regions; because the mean was not removed, this component can be interpreted as the mean pattern across the time steps. 
 The second component indicates a temperature differential between much of the Eastern hemisphere and the Americas and parts of Africa. 
 The third component indicates a temperature differential between Africa and Europe and other regions.
+It is plausible that the second and third components correspond to the regular diurnal cycle, which is likely one of the largest sources of regular variation in the one-year timespan considered here.
 While these components were extracted from a fairly short simulation run, the results demonstrate the ability of TributaryPCA to extract important climate patterns while running in a distributed in situ environment, without saving any raw data for post-processing.
 \begin{figure}
     \centering
