### 2024-06-17 #134 @ltang85
@simonbyrne @luraess The simulation is a whole year and it is in the beginning of sec 5.3. To clarify, we have also added one sentence into sec 5.3 - "It is plausible that the second and third components correspond to the regular diurnal cycle, which is likely one of the largest sources of regular variation in the one-year timespan considered here."
