### 2024-06-17 #134 @luraess
@ltang85 there some minor tweaks needed in order to be able to accept the paper:
- It would be useful to update the README content in your main repo https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale; currently, it contains information about preparing the proceeding submission. Ideally, it would contain a brief summary of the repo's content and its purpose (paper and scripts to reproduce results, which software it relates to)
- Further, we need a Zenodo DOI for the https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale repository, as well as a version number. This is needed for reproducibility purpose. You could link this repo with Zenodo, and tag a release which would then automatically generate a Zenodo archive with corresponding DOI (see https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)

Upon completing these 2 steps, we should be finally good to go
