+++ README.md
@@ -21,3 +21,7 @@ This research used resources provided by the Los Alamos National Laboratory Inst
 ## Disclaimer
 
 This paper is approved for public release under LA-UR-22-31718.
+
+## Contact
+
+Li Tang (ltang@lanl.gov)
