### 2024-06-18 #134 @luraess
@ltang85 we are soon ready to go. Could you please check
- the missing and invalid DOIs reported here: https://github.com/JuliaCon/proceedings-review/issues/134#issuecomment-2176914379
- and the final proof: https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00134/jcon.00134/10.21105.jcon.00134.pdf

Once I get your final feedback, we could proceed with accepting your manuscript - thanks!
