+++ paper/ref.bib
@@ -15,7 +15,7 @@
   author={Ge, Hong and Xu, Kai and Ghahramani, Zoubin},
   booktitle={International conference on artificial intelligence and statistics},
   pages={1682--1690},
-  doi={https://doi.org/10.17863/CAM.42246},
+  doi={10.17863/CAM.42246},
   year={2018},
   organization={PMLR}
 }
@@ -26,7 +26,7 @@
   journal={SoftwareX},
   volume={12},
   pages={100561},
-  doi={https://doi.org/10.1016/j.softx.2020.100561},
+  doi={10.1016/j.softx.2020.100561},
   year={2020},
   publisher={Elsevier}
 }
@@ -47,7 +47,7 @@
   volume={10},
   number={10},
   pages={2618--2644},
-  doi={https://doi.org/10.1029/2018MS001562},
+  doi={10.1029/2018MS001562},
   year={2018},
   publisher={Wiley Online Library}
 }
@@ -59,7 +59,7 @@
   volume={11},
   number={7},
   pages={2089--2129},
-  doi={https://doi.org/10.1029/2018MS001603},
+  doi={10.1029/2018MS001603},
   year={2019},
   publisher={Wiley Online Library}
 }
@@ -71,7 +71,7 @@
   volume={123},
   number={23},
   pages={13--046},
-  doi={https://doi.org/10.1029/2018JD028927},
+  doi={10.1029/2018JD028927},
   year={2018},
   publisher={Wiley Online Library}
 }
@@ -83,7 +83,7 @@
   volume={11},
   number={5},
   pages={1253--1268},
-  doi={https://doi.org/10.1029/2018MS001562},
+  doi={10.1029/2018MS001562},
   year={2019},
   publisher={Wiley Online Library}
 }
@@ -95,7 +95,7 @@
   volume={11},
   number={8},
   pages={2377--2411},
-  doi={ https://doi.org/10.1029/2019MS001629},
+  doi={10.1029/2019MS001629},
   year={2019},
   publisher={Wiley Online Library}
 }
@@ -145,7 +145,7 @@ url = "https://journals.ametsoc.org/view/journals/clim/20/3/jcli3996.1.xml"
 @BOOK{ssw_def,
 title = {Middle atmosphere dynamics},
 author = {Andrews, David  and Leovy, Conway  and Holton, James },
-doi = {https://doi.org/10.1038/294519a0},
+doi = {10.1038/294519a0},
 url = {https://www.osti.gov/biblio/5936274},
 place = {United States},
 year = {1987},
@@ -168,7 +168,7 @@ month = {1}
   author={Kelly, Christopher and Ha, Sungsoo and Huck, Kevin and Van Dam, Hubertus and Pouchard, Line and Matyasfalvi, Gyorgy and Tang, Li and D'Imperio, Nicholas and Xu, Wei and Yoo, Shinjae and others},
   booktitle={ISAV'20 In Situ Infrastructures for Enabling Extreme-Scale Analysis and Visualization},
   pages={15--19},
-  doi={https://doi.org/10.1145/3426462.3426465},
+  doi={10.1145/3426462.3426465},
   year={2020}
 }
 
@@ -192,7 +192,7 @@ journal = {Journal of Advances in Modeling Earth Systems},
 volume = {11},
 number = {7},
 pages = {2089-2129},
-doi = {https://doi.org/10.1029/2018MS001603},
+doi = {10.1029/2018MS001603},
 year = {2019}
 }
 
@@ -202,7 +202,7 @@ title = {{In Situ Methods, Infrastructures, and Applications on High Performance
[…]
