+++ paper/ref.bib
@@ -37,6 +37,7 @@
   booktitle={Proceedings of the JuliaCon Conferences},
   volume={1},
   pages={68},
+  doi={10.21105/jcon.00068},
   year={2021}
 }
 
@@ -531,10 +532,15 @@ publisher={Copernicus GmbH}
   publisher={SIAM}
 }
 
-@misc{scorpio,
-  title = {{SCORPIO}, A high-level Parallel I/O Library},
-  key = {SCORPIO},
-  howpublished = {\url{https://github.com/E3SM-Project/scorpio/}}
+
+@inproceedings{scorpio,
+  title={SCORPIO: a scalable two-phase parallel I/O library with application to a large scale subsurface simulator},
+  author={Sreepathi, Sarat and Sripathiy, Vamsi and Mills, Richard and Hammondz, Glenn and Mahinthakumar, G Kumar},
+  booktitle={20th Annual International Conference on High Performance Computing},
+  pages={443--451},
+  doi={10.1109/HiPC.2013.6799128},
+  year={2013},
+  organization={IEEE}
 }
 
 @inproceedings{jin2022accelerating,
