+++ paper/ref.bib
@@ -517,6 +517,7 @@ publisher={Copernicus GmbH}
   title={{AdaOja}: Adaptive learning rates for streaming {PCA}},
   author={Henriksen, Amelia and Ward, Rachel},
   journal={arXiv preprint arXiv:1905.12115},
+  doi={10.48550/arXiv.1905.12115},
   year={2019}
 }
 
