+++ paper/ref.bib
@@ -283,6 +283,17 @@ key = {SENSEI},
   year={2013}
 }
 
+@article{panda2021mvapich,
+  title={The MVAPICH project: Transforming research into high-performance MPI library for HPC community},
+  author={Panda, Dhabaleswar Kumar and Subramoni, Hari and Chu, Ching-Hsiang and Bayatpour, Mohammadreza},
+  journal={Journal of Computational Science},
+  volume={52},
+  pages={101208},
+  doi={10.1016/j.jocs.2020.101208},
+  year={2021},
+  publisher={Elsevier}
+}
+
 @inproceedings{gabriel2004open,
   title={Open {MPI}: Goals, concept, and design of a next generation {MPI} implementation},
   author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
