+++ paper/ref.bib
@@ -258,11 +258,17 @@ howpublished = {\url{https://sensei-insitu.org/}},
 key = {SENSEI},
 }
 
-@misc{gropp1996user,
-  title={User’s Guide for {MPICH}, a Portable Implementation of {MPI}},
-  author={Gropp, William and Lusk, Ewing},
+
+@article{gropp1996high,
+  title={A high-performance, portable implementation of the MPI message passing interface standard},
+  author={Gropp, William and Lusk, Ewing and Doss, Nathan and Skjellum, Anthony},
+  journal={Parallel computing},
+  volume={22},
+  number={6},
+  pages={789--828},
+  doi={10.1016/0167-8191(96)00024-5},
   year={1996},
-  publisher={Mathematics and Computer Science Division, Argonne National Laboratory~…}
+  publisher={Elsevier}
 }
 
 @article{walker1996mpi,
