+++ paper/ref.bib
@@ -258,6 +258,15 @@ howpublished = {\url{https://sensei-insitu.org/}},
 key = {SENSEI},
 }
 
+@inproceedings{ayachit2016sensei,
+  title={The SENSEI generic in situ interface},
+  author={Ayachit, Utkarsh and Whitlock, Brad and Wolf, Matthew and Loring, Burlen and Geveci, Berk and Lonie, David and Bethel, E Wes},
+  booktitle={2016 second workshop on in situ infrastructures for enabling extreme-scale Analysis and visualization (ISAV)},
+  pages={40--44},
+  doi={10.5555/3018859.3018867},
+  year={2016},
+  organization={IEEE}
+}
 
 @article{gropp1996high,
   title={A high-performance, portable implementation of the MPI message passing interface standard},
