+++ paper/paper.tex
@@ -98,7 +98,7 @@ Conventionally, large-scale HPC applications write their simulated scientific da
 
 \subsection{Related Work}
 
-The in situ approach has been studied in recent years and has been employed in different aspects of HPC. First, there are existing in situ frameworks, such as Ascent~\cite{ascent}, ParaView Catalyst~\cite{catalyst}, SENSEI~\cite{sensei_url}, VisIt libSIM~\cite{visit}, and SmartSim~\cite{chen2016smartsim}. These developed frameworks aim to help users build data analysis and data visualization, but none of them supports the Julia programming language. Second, the open-sourced ADIOS/ADIOS2~\cite{adios, godoy2020adios} data I/O system and the studies of Chimbuko/CODAR~\cite{kelly2020chimbuko} have explored data reduction for in situ approaches and researched the performance optimization of the whole in situ data analysis workflow. Last, previous studies have been made on optimizing the complex workflow that involves complex HPC applications and post-hoc data analysis. However, these studies still require significant I/O on disk storage and are not able to help reduce the data. For a more comprehensive summary of in situ infrastructures, readers are referred to~\cite{insitu_star,insitu_terminology}. 
+The in situ approach has been studied in recent years and has been employed in different aspects of HPC. First, there are existing in situ frameworks, such as Ascent~\cite{ascent}, ParaView Catalyst~\cite{catalyst}, SENSEI~\cite{ayachit2016sensei}, VisIt libSIM~\cite{visit}, and SmartSim~\cite{chen2016smartsim}. These developed frameworks aim to help users build data analysis and data visualization, but none of them supports the Julia programming language. Second, the open-sourced ADIOS/ADIOS2~\cite{adios, godoy2020adios} data I/O system and the studies of Chimbuko/CODAR~\cite{kelly2020chimbuko} have explored data reduction for in situ approaches and researched the performance optimization of the whole in situ data analysis workflow. Last, previous studies have been made on optimizing the complex workflow that involves complex HPC applications and post-hoc data analysis. However, these studies still require significant I/O on disk storage and are not able to help reduce the data. For a more comprehensive summary of in situ infrastructures, readers are referred to~\cite{insitu_star,insitu_terminology}. 
 
 
 The demand for scalable algorithms for analyzing climate data is growing rapidly as computational climate simulations scale up, leading to growing output data size and prompting scientists to turn to in situ analysis approaches. For example, in situ analysis of eddies in large-scale MPAS-Ocean simulation data was done by Woodring et al.~\cite{7192723}. In this work, we focus on two data analysis questions of interest to climate scientists and demonstrate their in situ application. For extreme weather event analysis, the connection of SSW with the polar vortex events is a focus area for climatologists~\cite{ssw_survey}. More recently, scientists have also used extreme-value distribution models for capturing rare climate phenomena and further analyzing them in detail~\cite{gev_huang2016estimating, gev_wang2017detection}. In our recent work, the usefulness of in situ SSW analysis was also demonstrated~\cite{ssw_isav}.
@@ -185,7 +185,7 @@ int* p_arr, int* p_len, MPI_Fint *Fcomm){
 \subsection{Julia Interface}
 
 The Julia interface mainly defines the format of input data, so all in situ Julia data analysis methods can be plugged-in without additional E3SM compilation, thanks to Julia's Just-in-time (JIT) compilation.
[…]
