### 2024-06-18 #134 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
