+++ paper/ref.bib
@@ -250,14 +250,6 @@ url = "https://journals.ametsoc.org/view/journals/clim/30/14/jcli-d-16-0465.1.xm
 author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Eric and Childs, Hank and Geveci, Berk and Harrison, Cyrus}, title = {The {ALPINE} In Situ Infrastructure: Ascending from the Ashes of Strawman}, year = {2017}, isbn = {9781450351393}, publisher = {Association for Computing Machinery}, address = {New York, NY, USA}, url = {10.1145/3144769.3144778}, doi = {10.1145/3144769.3144778}, booktitle = {Proceedings of the In Situ Infrastructures on Enabling Extreme-Scale Analysis and Visualization}, pages = {42–46}, numpages = {5}, keywords = {Scientific Visualization, HPC, In Situ}, location = {Denver, CO, USA}, series = {ISAV’17} }
 
 
-@Misc{sensei_url,
-title  = "{{SENSEI}:Scalable in situ analysis and visualization}",
-url    = "",
-year   = "2021 (accessed April 6, 2022)",
-howpublished = {\url{https://sensei-insitu.org/}},
-key = {SENSEI},
-}
-
 @inproceedings{ayachit2016sensei,
   title={The SENSEI generic in situ interface},
   author={Ayachit, Utkarsh and Whitlock, Brad and Wolf, Matthew and Loring, Burlen and Geveci, Berk and Lonie, David and Bethel, E Wes},
@@ -291,12 +283,6 @@ key = {SENSEI},
   publisher={ASFRA BV}
 }
 
-@inproceedings{panda2013mvapich,
-  title={The {MVAPICH} project: Evolution and sustainability of an open source production quality {MPI} library for {HPC}},
-  author={Panda, Dhabaleswar K and Tomko, Karen and Schulz, Karl and Majumdar, Amitava},
-  booktitle={Workshop on Sustainable Software for Science: Practice and Experiences, held in conjunction with Int’l Conference on Supercomputing (WSSPE)},
-  year={2013}
-}
 
 @article{panda2021mvapich,
   title={The MVAPICH project: Transforming research into high-performance MPI library for HPC community},
