+++ paper/ref.bib
@@ -489,19 +489,6 @@ doi={10.1145/3490138.3490142},
 numpages = {6}
 }
 
-
-@article{gev_bonsal2001characteristics,
-title={Characteristics of daily and extreme temperatures over Canada},
-author={Bonsal, BR and Zhang, X and Vincent, LA and Hogg, WD},
-journal={Journal of Climate},
-volume={14},
-number={9},
-pages={1959--1976},
-doi={10.1175/1520-0442(2001)014%3C1959:CODAET%3E2.0.CO;2},
-year={2001},
-publisher={American Meteorological Society}
-}
-
 @article{gev_huang2016estimating,
 title={Estimating changes in temperature extremes from millennial-scale climate simulations using generalized extreme value (GEV) distributions},
 author={Huang, Whitney K and Stein, Michael L and McInerney, David J and Sun, Shanshan and Moyer, Elisabeth J},
