+++ paper/ref.bib
@@ -250,14 +250,14 @@ url = "https://journals.ametsoc.org/view/journals/clim/30/14/jcli-d-16-0465.1.xm
 author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Eric and Childs, Hank and Geveci, Berk and Harrison, Cyrus}, title = {The {ALPINE} In Situ Infrastructure: Ascending from the Ashes of Strawman}, year = {2017}, isbn = {9781450351393}, publisher = {Association for Computing Machinery}, address = {New York, NY, USA}, url = {10.1145/3144769.3144778}, doi = {10.1145/3144769.3144778}, booktitle = {Proceedings of the In Situ Infrastructures on Enabling Extreme-Scale Analysis and Visualization}, pages = {42–46}, numpages = {5}, keywords = {Scientific Visualization, HPC, In Situ}, location = {Denver, CO, USA}, series = {ISAV’17} }
 
 
-@inproceedings{ayachit2016sensei,
-  title={The SENSEI generic in situ interface},
-  author={Ayachit, Utkarsh and Whitlock, Brad and Wolf, Matthew and Loring, Burlen and Geveci, Berk and Lonie, David and Bethel, E Wes},
-  booktitle={2016 second workshop on in situ infrastructures for enabling extreme-scale Analysis and visualization (ISAV)},
-  pages={40--44},
-  doi={10.5555/3018859.3018867},
-  year={2016},
-  organization={IEEE}
+@incollection{bethel2022sensei,
+  title={The sensei generic in situ interface: tool and processing portability at scale},
+  author={Bethel, E Wes and Loring, Burlen and Ayachit, Utkarsh and Camp, David and Duque, Earl PN and Ferrier, Nicola and Insley, Joseph and Gu, Junmin and Kress, James and O’Leary, Patrick and others},
+  booktitle={In Situ Visualization for Computational Science},
+  pages={281--306},
+  doi={10.1007/978-3-030-81627-8_13},
+  year={2022},
+  publisher={Springer}
 }
 
 @article{gropp1996high,
