### 2024-06-18 caa1c78
Update paper.tex
