+++ paper/ref.bib
@@ -300,7 +300,7 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
   author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
   booktitle={European Parallel Virtual Machine/Message Passing Interface Users’ Group Meeting},
   pages={97--104},
-  doi={10.1007/978-3-540-30218-6\_19},
+  doi={10.1007/978-3-540-30218-6_19},
   year={2004},
   organization={Springer}
 }
