+++ paper/ref.bib
@@ -508,7 +508,7 @@ publisher={Copernicus GmbH}
   author={Broderick, Tamara and Boyd, Nicholas and Wibisono, Andre and Wilson, Ashia C and Jordan, Michael I},
   journal={Advances in neural information processing systems},
   volume={26},
-  doi={https://dl.acm.org/doi/10.5555/2999792.2999805},
+  doi={10.5555/2999792.2999805},
   year={2013}
 }
 
