### 2024-06-19 #134 @luraess
@ltang85 There is compilation error now https://github.com/JuliaCon/proceedings-papers/actions/runs/9577082813/job/26404686944

Can you try checking with your local build and see why since the recent change (from here https://github.com/JuliaCon/proceedings-review/issues/134#issuecomment-2177070547) the PDF is no longer compiling. Thanks!
