+++ paper/ref.bib
@@ -255,7 +255,7 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
   author={Bethel, E Wes and Loring, Burlen and Ayachit, Utkarsh and Camp, David and Duque, Earl PN and Ferrier, Nicola and Insley, Joseph and Gu, Junmin and Kress, James and O’Leary, Patrick and others},
   booktitle={In Situ Visualization for Computational Science},
   pages={281--306},
-  doi={10.1007/978-3-030-81627-8_13},
+  doi={10.1007/978-3-030-81627-8\_13},
   year={2022},
   publisher={Springer}
 }
@@ -300,7 +300,7 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
   author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
   booktitle={European Parallel Virtual Machine/Message Passing Interface Users’ Group Meeting},
   pages={97--104},
-  doi={10.1007/978-3-540-30218-6_19},
+  doi={10.1007/978-3-540-30218-6\_19},
   year={2004},
   organization={Springer}
 }
