+++ paper/ref.bib
@@ -131,7 +131,7 @@
 
 @article {ssw,
 author = "Andrew J. Charlton and Lorenzo M. Polvani",
-title = "A New Look at Stratospheric Sudden Warmings. Part I: Climatology and Modeling Benchmarks",
+title = "A New Look at Stratospheric Sudden Warmings. {P}art {I}: Climatology and Modeling Benchmarks",
 journal = "Journal of Climate",
 year = "2007",
 publisher = "American Meteorological Society",
@@ -251,17 +251,17 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
 
 
 @incollection{bethel2022sensei,
-  title={The sensei generic in situ interface: tool and processing portability at scale},
+  title={The {SENSEI} generic in situ interface: tool and processing portability at scale},
   author={Bethel, E Wes and Loring, Burlen and Ayachit, Utkarsh and Camp, David and Duque, Earl PN and Ferrier, Nicola and Insley, Joseph and Gu, Junmin and Kress, James and O’Leary, Patrick and others},
   booktitle={In Situ Visualization for Computational Science},
   pages={281--306},
-  doi={10.1007/978-3-030-81627-8\_13},
+  doi={10.1007/978-3-030-81627-8\textunderscore13},
   year={2022},
   publisher={Springer}
 }
 
 @article{gropp1996high,
-  title={A high-performance, portable implementation of the MPI message passing interface standard},
+  title={A high-performance, portable implementation of the {MPI} message passing interface standard},
   author={Gropp, William and Lusk, Ewing and Doss, Nathan and Skjellum, Anthony},
   journal={Parallel computing},
   volume={22},
@@ -285,7 +285,7 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
 
 
 @article{panda2021mvapich,
-  title={The MVAPICH project: Transforming research into high-performance MPI library for HPC community},
+  title={The {MVAPICH} project: Transforming research into high-performance {MPI} library for {HPC} community},
   author={Panda, Dhabaleswar Kumar and Subramoni, Hari and Chu, Ching-Hsiang and Bayatpour, Mohammadreza},
   journal={Journal of Computational Science},
   volume={52},
@@ -300,7 +300,7 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
   author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
   booktitle={European Parallel Virtual Machine/Message Passing Interface Users’ Group Meeting},
   pages={97--104},
-  doi={10.1007/978-3-540-30218-6\_19},
+  doi={10.1007/978-3-540-30218-6\textunderscore19},
   year={2004},
   organization={Springer}
 }
@@ -351,7 +351,7 @@ numpages =  {9}
 
 @inproceedings{adios,
 author = {Lofstead, Jay F. and Klasky, Scott and Schwan, Karsten and Podhorszki, Norbert and Jin, Chen},
-title = {Flexible IO and Integration for Scientific Codes through the Adaptable IO System ({ADIOS})},
+title = {Flexible {IO} and Integration for Scientific Codes through the Adaptable IO System ({ADIOS})},
 year = {2008},
 isbn = {9781605581569},
 publisher = {Association for Computing Machinery},
@@ -431,7 +431,7 @@ numpages =  {10}
 
 @INPROCEEDINGS{8514879,
[…]
