### 2024-06-19 #134 @ltang85
> @luraess Done.

@luraess This includes a few changes for fixing the DOIs and refs after the proof read generation, do we need a new proof read generation?
