### 2024-06-19 531a55c
Update paper.tex
