+++ paper/ref.bib
@@ -295,14 +295,14 @@ author = {Larsen, Matthew and Ahrens, James and Ayachit, Utkarsh and Brugger, Er
   publisher={Elsevier}
 }
 
-@inproceedings{gabriel2004open,
-  title={Open {MPI}: Goals, concept, and design of a next generation {MPI} implementation},
-  author={Gabriel, Edgar and Fagg, Graham E and Bosilca, George and Angskun, Thara and Dongarra, Jack J and Squyres, Jeffrey M and Sahay, Vishal and Kambadur, Prabhanjan and Barrett, Brian and Lumsdaine, Andrew and others},
-  booktitle={European Parallel Virtual Machine/Message Passing Interface Users’ Group Meeting},
-  pages={97--104},
-  doi={10.1007/978-3-540-30218-6\textunderscore19},
-  year={2004},
-  organization={Springer}
+@inproceedings{graham2006open,
+  title={Open MPI: A high-performance, heterogeneous MPI},
+  author={Graham, Richard L and Shipman, Galen M and Barrett, Brian W and Castain, Ralph H and Bosilca, George and Lumsdaine, Andrew},
+  booktitle={2006 IEEE International Conference on Cluster Computing},
+  pages={1--9},
+  doi={10.1109/CLUSTR.2006.311904},
+  year={2006},
+  organization={IEEE}
 }
 
 
