+++ paper/header.tex
@@ -15,7 +15,7 @@
 \author[1]{Earl Lawrence}
 \affil[1]{Computer, Computational, and Statistical Sciences (CCS), Los Alamos National Laboratory}
 \affil[2]{Department of Computer Science and Engineering, Indian Institute of Technology Kanpur}
-\affil[3]{Department of Statistics, University of Michigan}
+\affil[3]{Google LLC}
 \affil[4]{Computational Science Initiative, Brookhaven National Laboratory}
 
 
