### 2024-06-20 #134 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
