### 2024-06-22 #134 @lucaferranti
@ltang85 I see e.g. in [this commit](https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale/commit/42b250364e8ab7c6b17c6c5ac73529f0e6ee9aeb) you have edited directly the `header.txt` file. Note that file is automatically generated and should not be edited manually. Could you update the insitutions in the [`paper.yml`](https://github.com/ltang85/In-Situ-data-Analysis-with-Julia-for-Climate-Simulations-at-Large-Scale/blob/main/paper/paper.yml#L33-L41) file? that should fix the issue
