### 2024-06-22 #134 @luraess
Thanks @lucaferranti for your input. @ltang85 please update the corresponding yml file and check with `@editorialbot generate pdf` whether all works as expected. I can then regenerate the final proof.
