### 2024-06-23 #134 @luraess
@ltang85 congratulation for having your paper published in the JuliaCon Proceedings! Thanks for your cooperation during the entire revision and publication process.
@williamfgc @simonbyrne thank you for helping out with the revisions!
👉  The paper is now available at https://doi.org/10.21105/jcon.00134
